split loaders to separate file
